extracted tutorial modal to new component
